Fix warnings and status flags in EOF to OK package conversion
If a client connects to ProxySQL not using CLIENT_DEPRECATE_EOF and it
caches a resultset, when a client using CLIENT_DEPRECATE_EOF executes
the same query it will get a resultset with wrong warnings and status
flags.

This is because warnings and status flags position must be swapped when
converting an EOF packet to an OK packet.

More info about warnings and status flags position in EOF and OK
packets:
https://dev.mysql.com/doc/dev/mysql-server/latest/page_protocol_basic_eof_packet.html
https://dev.mysql.com/doc/dev/mysql-server/latest/page_protocol_basic_ok_packet.html